% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToOBS.R
\name{connectToOBS}
\alias{connectToOBS}
\title{Create the R OBS Websocket object}
\usage{
connectToOBS(obs)
}
\arguments{
\item{obs}{an OBSwebsocket object}
}
\description{
Use an OBSWebsocket object to create a connection to the OBS Websocket server
The WebSocket server (Use "Tools" menu in OBS) must be activated.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
disconnectFromOBS(myobs)
}
\author{
FML
}
