% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-CreateInput.R
\name{CreateInput}
\alias{CreateInput}
\title{Create an input}
\usage{
CreateInput(
  obs = obs,
  sceneName,
  inputName,
  inputKind,
  inputSettings,
  sceneItemEnabled = TRUE,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The scene to which the input will be added}

\item{inputName}{The Name of the new input to be created}

\item{inputKind}{The kind of input to be created}

\item{inputSettings}{The settings of the new input. If missing the input will be created with default settings}

\item{sceneItemEnabled}{Whether to set the created scene item to enabled or disabled. Default TRUE}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Create an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
CreateInput(obs = myobs, sceneName = "Two cams", inputName = "New Browser", inputKind = "browser_source")
CreateInput(obs = myobs, sceneName = "Two cams", inputName = "New Browser B", inputKind = "browser_source", inputSettings = '{"height":602,"width":802}')
disconnectFromOBS(myobs)
}
}
\author{
FML
}
