% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWebsocket.R
\name{createOBSWebsocket}
\alias{createOBSWebsocket}
\title{Create the R OBS Websocket object}
\usage{
createOBSWebsocket(address = "ws://127.0.0.1", port = 4455)
}
\arguments{
\item{address}{Address to connect to the OBS websocket server (ws:// will be added if not provided) Default :127.0.0.1}

\item{port}{Port used for the connection. Default 4455}
}
\description{
Create a websocket object to connect to OBS and send commands.
The WebSocket server (Use "Tools" menu in OBS) must be activated.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
