% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigR-CreateSceneCollection.R
\name{CreateSceneCollection}
\alias{CreateSceneCollection}
\title{Create a new scene collection}
\usage{
CreateSceneCollection(
  obs = obs,
  sceneCollectionName = "New Collection",
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneCollectionName}{Name for the new scene collection. Default "New Collection"}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default TRUE}
}
\description{
Create a new scene collection and switch to it so you can directly store your new scenes
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
CreateSceneCollection(obs = myobs, sceneCollectionName="NewCollectionName")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
