% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter-CreateSourceFilter.R
\name{CreateSourceFilter}
\alias{CreateSourceFilter}
\title{Creates a new filter for a source}
\usage{
CreateSourceFilter(
  obs = obs,
  sourceName,
  filterName,
  filterKind,
  filterSettings,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sourceName}{The name of the target source to add the filter}

\item{filterName}{The name of the filter to create}

\item{filterKind}{The kind of the filter to create}

\item{filterSettings}{The settings of the new filter. If missing the filter will be created with default settings}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Creates a new filter for a source in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
CreateSourceFilter(obs = myobs, sourceName= "Audio 1", filterName = "MyLimiterFilter", filterKind = "limiter_filter", filterSettings = '{"release_time":50,"threshold":-5.0}')
disconnectFromOBS(myobs)
}
}
\author{
FML
}
