% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnectFromOBS.R
\name{disconnectFromOBS}
\alias{disconnectFromOBS}
\title{Create the R OBS Websocket object}
\usage{
disconnectFromOBS(obs = obs)
}
\arguments{
\item{obs}{an OBSwebsocket object}
}
\description{
disconnect from the OBS websocket server
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
