% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-GetCurrentProgramScene.R
\name{GetCurrentProgramScene}
\alias{GetCurrentProgramScene}
\title{Asks for the name of the current program scene}
\usage{
GetCurrentProgramScene(
  obs = obs,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestID}{The requestID used to provide feedback. Default UUIDgenerate()}
}
\description{
Get the name of the current program scene using an existing connection to OBS
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetCurrentProgramScene(obs = myobs)
disconnectFromOBS(myobs)

}
\author{
FML
}
