% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-GetGroupList.R
\name{GetGroupList}
\alias{GetGroupList}
\title{Gets the list of all groups}
\usage{
GetGroupList(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestID}{The requestID used to provide feedback. Default UUIDgenerate()}
}
\description{
Gets the list of all groups from OBS using an existing connection to OBS.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetGroupList(obs = myobs)
disconnectFromOBS(myobs)
}
\author{
FML
}
