% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR-GetHotkeyList.R
\name{GetHotkeyList}
\alias{GetHotkeyList}
\title{Gets the hot key list}
\usage{
GetHotkeyList(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the hto key list from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetHotkeyList(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
