% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-GetInputKindList.R
\name{GetInputKindList}
\alias{GetInputKindList}
\title{ask for a list of the input kinds}
\usage{
GetInputKindList(
  obs = obs,
  unversioned = FALSE,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{unversioned}{True: Return all kinds as unversioned, False: Return with version suffixes}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Get the list of the input kinds in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetInputKindList(obs = myobs)
GetInputKindList(obs = myobs, unversioned = TRUE)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
