% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-GetInputList.R
\name{GetInputList}
\alias{GetInputList}
\title{ask for a list of the inputs}
\usage{
GetInputList(
  obs = obs,
  inputKind = NULL,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputKind}{Restrict the array to only inputs of the specified kind}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Get the list of the inputs in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetInputList(obs = myobs)
GetInputList(obs = myobs, inputKind = "browser_source")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
