% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-GetInputPropertiesListPropertyItems.R
\name{GetInputPropertiesListPropertyItems}
\alias{GetInputPropertiesListPropertyItems}
\title{Gets the items of a list property for an input}
\usage{
GetInputPropertiesListPropertyItems(
  obs = obs,
  inputName,
  propertyName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the input}

\item{propertyName}{The name of the property to get the list of items from.}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the items of a list property for an input in OBS using an existing connection to OBS.
For example get the list of available display for display capture
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetInputPropertiesListPropertyItems(obs = myobs, inputName = "Display Capture 1", propertyName = "displays")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
