% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-GetInputVolume.R
\name{GetInputVolume}
\alias{GetInputVolume}
\title{Gets the volume of an input}
\usage{
GetInputVolume(
  obs = obs,
  inputName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the new input to get the settings from}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the volume of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetInputVolume(obs = myobs, inputName = "Audio 1")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
