% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIR-GetMonitorList.R
\name{GetMonitorList}
\alias{GetMonitorList}
\title{Ask for the monitor List}
\usage{
GetMonitorList(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the monitor list using an existing connection to OBS
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetMonitorList(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
