% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigR-GetSceneCollectionList.R
\name{GetSceneCollectionList}
\alias{GetSceneCollectionList}
\title{Get the list of all the scene collections}
\usage{
GetSceneCollectionList(
  obs = obs,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default TRUE}
}
\description{
Get the list of all the scene collections (use ViewLog() to get access to the list)
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSceneCollectionList(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
