% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-GetSceneSceneTransitionOverride.R
\name{GetSceneSceneTransitionOverride}
\alias{GetSceneSceneTransitionOverride}
\title{Gets the scene "transition override" for a scene.}
\usage{
GetSceneSceneTransitionOverride(
  obs = obs,
  sceneName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene (must exist in obs)}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the scene "transition override" for a scene. from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSceneSceneTransitionOverride(obs = myobs, scene = "Muti Cam Scene" )
disconnectFromOBS(myobs)
}
}
\author{
FML
}
