% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fiter-GetSourceFilterDefaultSettings.R
\name{GetSourceFilterDefaultSettings}
\alias{GetSourceFilterDefaultSettings}
\title{Gets the default settings for a filter kind}
\usage{
GetSourceFilterDefaultSettings(
  obs = obs,
  filterKind,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{filterKind}{TFilter kind to get the default settings for}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\value{
The request ID (can be used to access the websocket data using the parseById function )
}
\description{
Gets the default settings for a filter kind in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSourceFilterDefaultSettings(obs = myobs, filterKind = "limiter_filter")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
