% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fiter-GetSourceFilterList.R
\name{GetSourceFilterList}
\alias{GetSourceFilterList}
\title{Get all the filters of a source}
\usage{
GetSourceFilterList(
  obs = obs,
  sourceName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sourceName}{The Name of the source to get the filtters from}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\value{
The request ID (can be used to access the websocket data using the parseById function )
}
\description{
Get all the filters of a source in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSourceFilterList(obs = myobs, sourceName = "Audio 1")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
