% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-GetSpecialInputs.R
\name{GetSpecialInputs}
\alias{GetSpecialInputs}
\title{Ask for a list of the special inputs}
\usage{
GetSpecialInputs(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Get the list of the special inputs in OBS using an existing connection to OBS.
(Desktop audio, mic...)
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSpecialInputs(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
