% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR-GetStats.R
\name{GetStats}
\alias{GetStats}
\title{Gets statistics about OBS, obs-websocket, and the current session}
\usage{
GetStats(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets statistics about OBS, obs-websocket, and the current session from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetStats(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
