% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR-GetStreamServiceSettings.R
\name{GetStreamServiceSettings}
\alias{GetStreamServiceSettings}
\title{Gets the stream service settings}
\usage{
GetStreamServiceSettings(
  obs = obs,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the stream service settings from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetStreamServiceSettings(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
