% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR-GetVersion.R
\name{GetVersion}
\alias{GetVersion}
\title{Gets the version of the currentplugin and rpc version}
\usage{
GetVersion(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Gets the version of the currentplugin and rpc version from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetVersion(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
