% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyToOBS.R
\name{identifyToOBS}
\alias{identifyToOBS}
\title{Create the R OBS Websocket object}
\usage{
identifyToOBS(obs = obs, password, eventSub = "33")
}
\arguments{
\item{obs}{an OBSwebsocket object}

\item{password}{the password of the OBS websocket server if authentication has been activated (strongly advisable)}

\item{eventSub}{Default 33}
}
\description{
Use an OBSWebsocket object already connected with connectToOBS() to the OBS Websocket server to identify the client
If authentication is on the OBS websocket server is activated, the password must provided
The WebSocket server (Use "Tools" menu in OBS) must be activated.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
