% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onRobs.R
\docType{package}
\name{onRobs-package}
\alias{onRobs}
\alias{onRobs-package}
\title{onRobs: R client for obs-websocket}
\description{
onRobs uses OBS Websockets to control OBS from R
To learn more about onRobs, start with the vignettes:
`browseVignettes(package = "onRobs")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franklentz/onRobs}
  \item \url{https://franklentz.github.io/onRobs/}
  \item Report bugs at \url{https://github.com/franklentz/onRobs/issues}
}

}
\keyword{internal}
