% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIR-OpenInputInteractDialog.R
\name{OpenInputInteractDialog}
\alias{OpenInputInteractDialog}
\title{Open the input interact Dialog box}
\usage{
OpenInputInteractDialog(
  obs = obs,
  inputName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{An obs websocket with an active connection. Default obs}

\item{inputName}{Name of the input to open the filters dialog of}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Open the input interact dialog box using an existing connection to OBS
(Use with browser ...)
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
OpenInputInteractDialog(obs = myobs, inputName = "Audio Capture 1")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
