% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseById.R
\name{parseById}
\alias{parseById}
\title{Use the ID of a request response to parse the response data}
\usage{
parseById(obs = obs, requestId)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId from which to get the data}
}
\value{
According to the request type corresponding to the id : a boolean, an integer, a string or a dataframe
}
\description{
Get function display the response message from the server but return the request id
This function use the request id to extract and parse the response data from the log of the robs websocket object
This is NOT yet fully functional
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")

disconnectFromOBS(myobs)
}
}
\author{
FML
}
