% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-PressInputPropertiesButton.R
\name{PressInputPropertiesButton}
\alias{PressInputPropertiesButton}
\title{Press a button in the properties of an input}
\usage{
PressInputPropertiesButton(
  obs = obs,
  inputName,
  propertyName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the input}

\item{propertyName}{The name of the button property to press    .}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Press a button in the properties of an input in OBS using an existing connection to OBS.
For example to refresh a browser input, use "refreshnocache".
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
PressInputPropertiesButton(obs = myobs, inputName = "BrowserA", propertyName = "refreshnocache")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
