% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-RemoveInput.R
\name{RemoveInput}
\alias{RemoveInput}
\title{Remove an input}
\usage{
RemoveInput(
  obs = obs,
  inputName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the new input to be removed}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Remove an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
RemoveInput(obs = myobs, inputName = "New Browser")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
