% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter-RemoveSourceFilter.R
\name{RemoveSourceFilter}
\alias{RemoveSourceFilter}
\title{Removes a filter for a source}
\usage{
RemoveSourceFilter(
  obs = obs,
  sourceName,
  filterName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sourceName}{The name of the target source to add the filter}

\item{filterName}{The name of the filter to create}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Removes a filter for a source in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
RemoveSourceFilter(obs = myobs, sourceName= "Audio 1", filterName = "MyLimiterFilter")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
