% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBS.R
\docType{package}
\name{ROBS-package}
\alias{ROBS}
\alias{ROBS-package}
\title{ROBS: R client for obs-websocket}
\description{
ROBS uses OBS Websockets to control OBS from R
To learn more about ROBS, start with the vignettes:
`browseVignettes(package = "ROBS")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franklentz/ROBS}
  \item \url{https://franklentz.github.io/ROBS/}
  \item Report bugs at \url{https://github.com/franklentz/ROBS/issues}
}

}
\keyword{internal}
