% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamR-SendStreamCaption.R
\name{SendStreamCaption}
\alias{SendStreamCaption}
\title{Sends CEA-608 caption text over the stream output}
\usage{
SendStreamCaption(
  obs = obs,
  captionText = "Test Caption",
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{captionText}{The text of the caption}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Sends CEA-608 caption text over the stream output using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SendStreamCaption(obs = myobs, captionText="My caption")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
