% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-SetCurrentProgramScene.R
\name{SetCurrentProgramScene}
\alias{SetCurrentProgramScene}
\title{Changes the Scene of the current program}
\usage{
SetCurrentProgramScene(
  obs = obs,
  sceneName = "main",
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene to switch to (must exist in obs)}

\item{requestID}{The requestID used to provide feedback. Default UUIDgenerate()}
}
\description{
Changes the Scene of the current program using an existing connection to OBS.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetCurrentProgramScene(obs = myobs, scene = "Muti Cam Scene")
SetCurrentProgramScene(obs = myobs, scene = "BRB")
disconnectFromOBS(myobs)
}
\author{
FML
}
