% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigR-SetCurrentSceneCollection.R
\name{SetCurrentSceneCollection}
\alias{SetCurrentSceneCollection}
\title{Switch the current scene collection}
\usage{
SetCurrentSceneCollection(
  obs = obs,
  sceneCollectionName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneCollectionName}{Name for the scene collection to switch to.}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default TRUE}
}
\description{
Switches to the specified scene collection.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetCurrentSceneCollection(obs = myobs, sceneCollectionName="Main")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
