% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputAudioBalance.R
\name{SetInputAudioBalance}
\alias{SetInputAudioBalance}
\title{Set the audio balance of an input}
\usage{
SetInputAudioBalance(
  obs = obs,
  inputName,
  inputAudioBalance = NULL,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the input to change the volume of}

\item{inputAudioBalance}{inputAudioBalance setting between 0 and 1}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Set the audio balance of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputAudioBalance(obs = myobs, inputName = "Audio 2", inputAudioBalance = 0.5)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
