% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputAudioMonitorType.R
\name{SetInputAudioMonitorType}
\alias{SetInputAudioMonitorType}
\title{Change the audio monitor type of an input}
\usage{
SetInputAudioMonitorType(
  obs = obs,
  inputName,
  monitorType,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The name of the input}

\item{monitorType}{The new monitorType for the input (OBS_MONITORING_TYPE_NONE, OBS_MONITORING_TYPE_MONITOR_ONLY,OBS_MONITORING_TYPE_MONITOR_AND_OUTPUT}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Change the Change the audio monitor type of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputAudioMonitorType(obs = myobs, inputName = "Audio 1", monitorType = "OBS_MONITORING_TYPE_MONITOR_ONLY")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
