% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputAudioSyncOffset.R
\name{SetInputAudioSyncOffset}
\alias{SetInputAudioSyncOffset}
\title{Set the audio balance of an input}
\usage{
SetInputAudioSyncOffset(
  obs = obs,
  inputName,
  inputAudioSyncOffset = NULL,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the input to change the volume of}

\item{inputAudioSyncOffset}{inputAudioBalance setting between -950 and 2000}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Set the audio balance of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputAudioSyncOffset(obs = myobs, inputName = "Audio 2", SetInputAudioSyncOffset = 1000)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
