% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputAudioTracks.R
\name{SetInputAudioTracks}
\alias{SetInputAudioTracks}
\title{Sets the settings of audio tracks for an input}
\usage{
SetInputAudioTracks(
  obs = obs,
  inputName,
  inputAudioTracks,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the input}

\item{inputAudioTracks}{The settings of the Audio Tracks.}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Sets the settings of audio tracks for an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputAudioTracks(obs = myobs, inputName = "Audio 1", inputAudioTracks = '{"1":true,"2":true,"3":true}')
disconnectFromOBS(myobs)
}
}
\author{
FML
}
