% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputMute.R
\name{SetInputMute}
\alias{SetInputMute}
\title{Mute or unmute an input}
\usage{
SetInputMute(
  obs = obs,
  inputName,
  inputMuted = TRUE,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the new input to be muted or unmuted}

\item{inputMuted}{Whether to mute the input or not. Default TRUE}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Sets the audio mute state of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputMute(obs = myobs, inputName = "Audio 2", inputMuted = FALSE)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
