% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputName.R
\name{SetInputName}
\alias{SetInputName}
\title{Change the name of an input}
\usage{
SetInputName(
  obs = obs,
  inputName,
  newInputName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The name of the new input to be changed}

\item{newInputName}{The new name for the input}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Change the name of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputName(obs = myobs, inputName = "New Browser", newInputName = "Main Browser")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
