% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputSettings.R
\name{SetInputSettings}
\alias{SetInputSettings}
\title{Sets the settings of an input}
\usage{
SetInputSettings(
  obs = obs,
  inputName,
  inputSettings,
  overlay = TRUE,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the new input to be created}

\item{inputSettings}{The settings of the new input. If missing the input will be created with default settings}

\item{overlay}{Apply the settings on top of existing ones if TRUE or reset the input to its defaults before applying settings if FALSE. Default TRUE}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Sets the settings of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputSettings(obs = myobs, inputName = "New Browser", inputsSettings = "{"height":602,"width":802}")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
