% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR-SetInputVolume.R
\name{SetInputVolume}
\alias{SetInputVolume}
\title{Set the volume of an input}
\usage{
SetInputVolume(
  obs = obs,
  inputName,
  inputVolumeMul = NULL,
  inputVolumeDb = NULL,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{inputName}{The Name of the input to change the volume of}

\item{inputVolumeMul}{Volume setting in mul between 0 and 20}

\item{inputVolumeDb}{Volume setting in dB between -100 and 26}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Set the volume of an input in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetInputVolume(obs = myobs, inputName = "Audio 2", inputVolumeMul = 10)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
