% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-SetSceneName.R
\name{SetSceneName}
\alias{SetSceneName}
\title{Sets the name of a scene (rename).}
\usage{
SetSceneName(
  obs = obs,
  sceneName,
  newSceneName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene to set the name (must exist in obs)}

\item{newSceneName}{The new name of the scene to set the name to}

\item{requestID}{The requestID used to provide feedback. Default UUIDgenerate()}
}
\description{
Sets the name of a scene (rename) from OBS using an existing connection to OBS.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetSceneName(obs = myobs, scene = "Muti Cam Scene")
SetSceneName(obs = myobs, scene = "BRB")
disconnectFromOBS(myobs)
}
\author{
FML
}
