% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-SetSceneSceneTransitionOverride.R
\name{SetSceneSceneTransitionOverride}
\alias{SetSceneSceneTransitionOverride}
\title{Sets the scene "transition override" for a scene.}
\usage{
SetSceneSceneTransitionOverride(
  obs = obs,
  sceneName,
  transitionName,
  transitionDuration = 300,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene (must exist in obs)}

\item{transitionName}{Name of the scene transition to use as override. Specify null to remove}

\item{transitionDuration}{Duration to use for any overridden transition. Specify null to remove. A number >= 50, <= 20000.}

\item{requestID}{The requestID used to provide feedback. Default UUIDgenerate()}
}
\description{
Sets the scene "transition override" for a scene from OBS using an existing connection to OBS.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetSceneSceneTransitionOverride(obs = myobs, scene = "Muti Cam Scene", transitionName = "a name", transitionDuration = 500)
disconnectFromOBS(myobs)
}
\author{
FML
}
