% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetStudioModeEnabled.R
\name{SetStudioModeEnabled}
\alias{SetStudioModeEnabled}
\title{Activate or desactivate the studio mode}
\usage{
SetStudioModeEnabled(
  obs = obs,
  studioModeEnabled = TRUE,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{studioModeEnabled}{TRUE (enabled) or FALSE (disabled). Default: TRUE}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Set whether studio mode is enabled or not on OBS using an existing connection to OBS
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetStudioModeEnabled(obs = myobs, studioModeEnabled = TRUE)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
