% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamR-StopStream.R
\name{StopStream}
\alias{StopStream}
\title{Stop the stream output}
\usage{
StopStream(obs = obs, requestId = uuid::UUIDgenerate(), verbose = TRUE)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Stop the stream output in OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
StopStream(obs = myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
