% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewLog.R
\name{ViewLog}
\alias{ViewLog}
\title{View the log of requests and the responses}
\usage{
ViewLog(obs)
}
\arguments{
\item{obs}{an OBSwebsocket object}
}
\value{
a data frame with the status of the request and the returned values
}
\description{
Use this to access to the values returned by your requests
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSceneSceneTransitionOverride(obs = myobs, scene = "Muti Cam Scene" )
ViewLog(myobs)
disconnectFromOBS(myobs)
}
}
\author{
FML
}
