% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-CreateScene.R
\name{CreateScene}
\alias{CreateScene}
\title{Create a new scene}
\usage{
CreateScene(
  obs = obs,
  sceneName = "New",
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{sceneName}{The name of the scene to switch to create}
}
\description{
Create a new scene in the current collection in OBS.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
CreateScene(obs = myobs, sceneName = "New Scene")
disconnectFromOBS(myobs)
}
\author{
FML
}
