% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneIR-GetSceneItemList.R
\name{GetSceneItemList}
\alias{GetSceneItemList}
\title{ask for a list of the items}
\usage{
GetSceneItemList(
  obs = obs,
  sceneName = "main",
  verbose = TRUE,
  requestId = uuid::UUIDgenerate()
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene to switch to (must exist in obs)}

\item{requestID}{The requestID used to provide feedback. Default UUIDgenerate()}
}
\description{
Changes the Scene of the current program using an existing connection to OBS.
}
\examples{

myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
GetSceneItemList(obs = myobs, sceneName = "BRB")
disconnectFromOBS(myobs)
}
\author{
FML
}
