% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-RemoveScene.R
\name{RemoveScene}
\alias{RemoveScene}
\title{Remove a Scene from OBS}
\usage{
RemoveScene(
  obs = obs,
  sceneName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene to remove (must exist in obs)}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Remove a Scene from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
RemoveScene(obs = myobs, sceneName = "Muti Cam Scene")
RemoveScene(obs = myobs, sceneName = "BRB")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
