% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SceneR-SetSceneName.R
\name{SetSceneName}
\alias{SetSceneName}
\title{Sets the name of a scene (rename).}
\usage{
SetSceneName(
  obs = obs,
  sceneName,
  newSceneName,
  requestId = uuid::UUIDgenerate(),
  verbose = TRUE
)
}
\arguments{
\item{obs}{an obs websocket with an active connection. Default obs}

\item{sceneName}{The name of the scene to set the name (must exist in obs)}

\item{newSceneName}{The new name of the scene to set the name to}

\item{requestId}{The requestId used to provide feedback. Default UUIDgenerate()}

\item{verbose}{More messages and outputs if TRUE. Default: TRUE}
}
\description{
Sets the name of a scene (rename) from OBS using an existing connection to OBS.
}
\examples{
\dontrun{
myobs <- createOBSWebsocket()
connectToOBS(obs = myobs)
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
SetSceneName(obs = myobs, sceneName = "Muti Cam Scene", newSceneName = "SC MultiCam")
disconnectFromOBS(myobs)
}
}
\author{
FML
}
